move convert_filter to misc
